requirements.txt
# Core Dependencies
agno==1.7.6
google-genai
openai>=1.0.0
python-dotenv==1.0.0
pydantic==2.8.2
pydantic-settings==2.3.4

# API Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx
aiohttp>=3.9.0

# Database
supabase
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0

# Cache and Queue
redis==5.0.1
aioredis==2.0.1
celery==5.3.4

# Retry and Circuit Breaker
tenacity==8.2.3
backoff==2.2.1

# Utilities
colorama==0.4.6
rich==13.7.1
asyncio==3.4.3
pytz==2024.1

# Development
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.12.0
black==24.4.2
flake8==7.1.0

# Logging
loguru==0.7.2

# Image Processing (for multimodal)
Pillow==10.3.0
python-magic>=0.4.27
# python-magic-bin>=0.4.14  # Apenas para Windows, comentado para Linux

# PDF Processing
pdf2image>=1.16.3  # Para conversão de PDF para imagem

# PDF Generation (for testing)
reportlab==4.0.9

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==2.18.0

# Google Calendar Integration
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0